+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.6.0][] (2022-05-04)
+
 New:
 
 - Add `roles`, `shared_link` and `contact_id` to the recipient in the document details response:
@@ -195,7 +197,8 @@ Fixes:
 
 - Initial release
 
-[Unreleased]: https://github.com/opti/panda_doc/compare/v0.5.3...HEAD
+[Unreleased]: https://github.com/opti/panda_doc/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/opti/panda_doc/compare/v0.5.3...v0.6.0
 [0.5.3]: https://github.com/opti/panda_doc/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/opti/panda_doc/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/opti/panda_doc/compare/v0.5.0...v0.5.1